controllers: Instantiate ViewHelper in each controller
BaseController does not really depend on View helper so it the helper
is only instantiated there to save us some typing. But if we want to inject
it via dependency injection container, we would need to pass it in each
controller’s constructor to BaseController anyway. So we can as well move
the instantiation there.