Use DI container for injecting View helper
Instead of instantiating its dependencies in each class, we are going
to switch to dependency injection. We will declare the dependencies using
typehinted constructor arguments and pass already instantiated dependency
objects to the constructor. This will loosen the coupling between classes
and allow us to pass mock objects for testing.

Rather than creating and pasing instances of dependencies manually,
we will let Dice dependency injection container library instantiate them
for us. We are setting CONTAINER value in fatfree’s Hive to a function
wrapping around Dice->create method (Dice does not support PSR-11),
which will make fatfree ask Dice to instantiate controllers.
Dice will recursively instantiate all of the controller’s dependencies
and then instantiate the controller itself.

For now, we are only injecting View helpers.
Also the controllers are often instantiating other controllers, including
themselves. This is a bad design but we are not going to address it here.